plot1d.py split
